+++ changelog.md
@@ -35,6 +35,8 @@
 - Added special named flags to the Creative Tabs
 - Added Pipe junctions to pipe world gen
 - Added new baby piranha plant model for vanilla consistency
+- Added Piranha Plant spawn animation
+- Improved Piranha Plant spawning, spawns during the day in Jungles
 - Increased mini goomba spawn rate & group size
 - Goombas car rarely spawn as a goombella variant
 - Fixed teleporting using Clear Pipes with Sable